More configurability to the footer
